Remove website-specific items to abc-website.
