+++ delivery/frontend/Gemfile
@@ -40,5 +40,6 @@ group :development do
   gem 'kramdown'
 end
 
+gem 'abc', :path => '../../'
 gem 'abc-core', :path => '../../core'
 gem 'abc-adapters', :path => '../../adapters'
