--- delivery/frontend/app/conductors/abc/frontend/conductors/shows_email_signup_form.rb
@@ -1,15 +1,0 @@
-require_relative 'base'
-
-module Abc
-  module Frontend
-    module Conductors
-      class ShowsEmailSignupForm < Base
-
-        def to_response
-          {:page => OpenStruct.new(:title => "Sign up to receive a notification")}
-        end
-
-      end
-    end
-  end
-end
