--- delivery/frontend/app/views/abc/frontend/email_signups/thank_you.html.haml
@@ -1,4 +1,0 @@
-%h1 Thank you!
-%p
-  We appreciate your interest in the project.
-  (You signed up as #{@data.email_signup.email}. If that's incorrect or changes, let us know!)
