+++ lib/abc.rb
@@ -1,4 +1,5 @@
 # Abc is the Afterburner CMS namespace.
-
 module Abc
 end
+
+require 'abc/base_conductor'
